pyproject.toml
[tool.poetry]
name = "iconic-chat-bot"
version = "0.1.0"
description = ""
authors = ["nhao2003 <nhathaodev@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.3"
python-dotenv = "^1.0.1"
flask = "^3.1.0"
google-generativeai = "^0.8.3"
pymongo = "^4.10.1"
sentence-transformers = "^3.3.1"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
